[2/10] vector->dynarray in MSXCommandEvent
MSXCommandEvent internally stores a 'vector<TclObject>', this patch
changes that to 'dynarray<TclObject>'.

See the documentation in dynarray.hh for details on that class. Summary:
dynarray is a dynamically allocated array, but once constructed it can no
longer change size. The only advantage of dynarray over vector is that
the dynarray object itself is smaller than vector (16 vs 24 bytes).

The real motivation for this patch is to reduce the size of the
MSXCommandEvent objecti (from 40 to 32 bytes). This will become
important later in this patch series.