[3/10] refactor StateChange creation in DiskChanger
After this patch all StateChange objects are now created using the
following pattern:
  stateChangeDistributor.distributeNew(make_shared<...>(...));
So 'make_shared' immediately followed by 'distributeNew()'. Later in
this series we'll exploit this by moving the construction inside the
'distributeNew()' function.

This patch also avoids some redundant 'TclObject <-> string'
conversions. But that's merely a bonus, not a goal, of this patch.