[4/10] StateChangeRecorder no longer inherits from StateChangeListener
Emulation is (or should be) 100% deterministic. This means that if we
start from a known state, and also know all input events (e.g. MSX
keyboard presses) then we can perfectly replay emulation and get to the
same resulting state.

Managing these input events is done via 'StateChange' objects. We have
producers of these objects. For example the keyboard code translates
host keyboard events to MSX-keyboard-matrix stateChange events. We also
have consumers of these events. For example (also) the keyboard
emulation code listens for keyboard-matrix-change events and then
updates the MSX keyboard matrix accordingly. Note that the same events
can also be produced during replay, so the MSX keyboard reacts the same
during 'live-play' and during 'replay'.

During 'live-play' all state-change-events need to be recorded (to be
able to be replayed later). Before this patch, the recorder was
implemented as an event-listener. But instead of reacting to (one kind
of) these events, it stores (all) the events in some log. This patch
changes that: the recorder is no longer a listener (it no longer
inherits from StateChangeListener), instead it is it's own category (it
inherits from StateChangeRecorder)

One one hand this complicates things: the state-change-distributor now
has to explicitly send the event to the recorder (before this was
implicit because the recorder was just another listener). On the other
hand this simplifies things because during replay the recorder no longer
"hears an echo" of what was already recorded.

But the main motivation for this patch is as preparation for later
patches in this series: we want to change how ownership of the
StateChange objects is handled. More specifically instead of letting the
recorder store (copies of) the events, we want to make the recorder
create the events. And that's not compatible with having the recorder
'just another listener'.