[5/10] change distributeNew() interface
Before callers of this function looked like:
    stateChangeDistributor.distributeNew(std::make_shared<KeyMatrixState>(
        time, row, press, release));
This patch changes that to:
    stateChangeDistributor.distributeNew<KeyMatrixState>(
        time, row, press, release);
In other words: instead of passing (a shared_ptr to) a StateChange
object to distributeNew(), we pass the type and the constructor
parameters of the to-be created StateChange object.

In the current patch distributeNew() still acts mostly the same as
before: it internally creates the shared_ptr and then processes it the
same as before. But in later patches this will be optimized further.

This patch also changes the order of the MSXCommandEvent constructor
parameters. Now all StateChange objects have 'time' as their first
parameter.