[6/10] remove StateChangeRecorder abstract base class
As preparation for the next patch, because that one will change the
current record() method into a templatized method. And template methods
cannot be virtual.

There was currently exactly one subclass of 'StateChangeRecorder' namely
'ReverseManager'. So the base class was not really needed. The only
consequence of removing it is that now 'StateChangeDistributor' needs to
work with a concrete class (ReverseManager) rather than an abstract
class (StateChangeRecorder) and this adds an extra #include dependency.