[7/10] change ownership handling of StateChange objects
At the start of this patch series they were handled via shared_ptr with
no clear (single) owner. After this patch the ReverseManager is the
clear (single) owner of these objects. And the objects are now stored by
unique_ptr.

Detail: in case the reverse-subsystem is not enabled, there is no
ReverseManager. And then the StateChangeDistributor creates temporary
StateChange objects on the stack (so with lifetime limited to the scope
of the function).

A consequence of this is that the interface of some functions changed.
For example, before:
    void signalStateChange(const std::shared_ptr<StateChange>& event)
after:
    void signalStateChange(const StateChange& event)
In fact the majority of this patch makes such mechanical changes.