Fix(?) race-condition in 8b7ca958f
The symptom is: openMSX occasionally stops with an error like:
   Tcl_AppendLimitedToObj called with shared object
when launched from catapult (actually when there's communication between
openMSX and catapult).

Commit 8b7ca958f changed 'std::string' to 'TclObject' in 'FileDropEvent'
and 'CliCommandEvent'. The only reason for this was to reduce the size
of these objects because they were the largest objects in 'EventVariant'
(sizeof(std::string)==32, sizeof(TclObject)==8).

Unfortunately this change introduced a race-condition because TclObject
is internally reference counted and that reference-count, for the
constructor parameter of CliCommandEvent, could get manipulated in both
the main- and the CliConnection-thread.

Solved by changing 'TclObject' to 'StringStorage' (is basically a
'char*', but with ownership semantics). Constructing/passing a
'StringStorage' over thread-boundaries should be safe. And also
sizeof(StringStorage)==8.