[8/10] handle std::variant in serialization
Our serialization framework could already handle saving/restoring of
polymorphic objects that are implemented via (sub)classes in a class
hierarchy. Another way to implement polymorphism is via std::variant.
This patch extends to serialization framework to handle std::variant.
Moreover it does so in a way that results in an identical looking
(textual) serialized form.

For example a replay file (.omr file) contains a list of state-change
events. When you rename .omr to .xml.gz you can examine it. The list may
contain stuff like this:
    <events>
      <item id="1706" type="KeyMatrixState">
        <StateChange>
          <time>
            <time>135452575200</time>
          </time>
        </StateChange>
        <row>8</row>
        <press>1</press>
        <release>0</release>
      </item>
      ...
    </events>
That is each item in the list has a 'type' attribute, the value of that
attribute determines the structure of the stuff inside the 'item' tag.

This example was from an old .omr file I had lying around. At the time
StateChange was a polymorphic class hierarchy. The next patch in this
series will instead model StateChange via std::variant. But this change
in internal representation has no influence on the serialized textual
output. This way there's no backwards compatibility problem with the
format of replay files.