[9/10] Store StateChange in std::variant
The previous patches in this series changed how ownership of StateChange
objects is handled: it's changed from 'vector<shared_ptr<StateChange>>'
to 'vector<unique_ptr<StateChange>>'.

Whenever you have a (vector of) unique_ptr's, you can ask yourself "Why
have this indirection at all?". In other words: "Why not store the
objects by value?". Often the answer is: "Because the type of the
pointer is (an abstract) base-class and there can be multiple concrete
sub-classes. That's also the case here.

However the previous patches in this series also reduced the maximum
size of these objects. Before the maximum size was 56 bytes, now it's
only 32 bytes. (Actually 24 bytes + vptr, this patch looses the vptr (8
bytes) but replaces that with a variant-index (probably less than 8
bytes, but including padding the total size is still 32)).

When the maximum object size is small (and/or when the size difference
between the largest and the smallest object is small) it's possible to
replace heap-allocated polymorphic objects with a std::variant. And
that's what this patch does.

Before we required:
- 16 bytes for the shared_ptr object itself.
- Between 16 and 56 bytes for the StateChange object itself (most common
  was 24 bytes).
- 16(?) bytes for the shared_ptr administration overhead (e.g. reference
  counter and weak_ptr reference counter)
- (And possibly some heap-administration overhead.)
So between 48 and 88 bytes, most common 56 bytes.
Now we require exactly 32 bytes for the std::variant.

So we require less memory and have one less indirection. Though the main
motivation for this patch series was to remove one of the last sources
of recurring heap-allocations. That is, once openMSX is fully started
and actual emulation is running, we no longer require constant heap
allocations (of small objects). Note: this is only true for the openMSX
code itself. Apparently my openGL graphics driver still performs lots of
small memory allocations per frame.

Obviously when the StateChange-log grows larger and larger we need to
allocate memory for it. But that memory gets allocated in large chunks,
not one (small) allocation per object.