[10/10] replace vector with deque
The previous commit message explained how the StateChange-log
(obviously) grows when the replay gets longer.

Before this log was stored in a std::vector, this patch changes that to a
std::deque. The difference is that vector internally uses one big chunk
of memory while deque is internally split in large chunks (e.g. each
containing 512 elements). So when constantly adding new elements,
occasionally vector needs to allocate a new memory block and copy all
existing elements to the new block. But deque only needs to allocate a
new chunk and start filling that.

This patch also (re-)enables serialization of std::deque and fixes a bug
in it (apparently we never needed this before).