[1/45] Replace class hierarchy with std::function
This is a long patch series. Each patch only makes a small tweak, but
combined this whole series causes +700 fewer heap-memory allocations
during openMSX startup.

Although there are +700 fewer allocations, the total allocated memory is
roughly the same. That is because e.g. two smaller allocations are
replaced with one bigger allocation (sum of the two old sizes is roughly
the same as the new size).

--

This specific patch replaces a class hierarchy like this:
   struct PolymorphicSaverBase {
      virtual void save(Archive&, void*) = 0;
   };
   ...
   std::vector<PolymorphicSaverBase*> savers;
with
   std::vector<std::function<void(Archive&, void*)>> savers;

That is: we had a base class with exactly 1 virtual function and a
vector of (pointers to) derived objects of this class. So the idea is to
have a collection of different implementation of this function. But this
is exactly the functionality that std::function<> offers.

Before we had many small objects (each separately allocated on the
heap). Each contained a vptr and a char* (16 bytes + possible heap
administration overhead). We also had a vector of pointers (8 bytes per
element).

Now we have a vector of std::function objects, these are initialized
with lambda functions. In the libstdc++ implementation each
std::function object takes 32 bytes. Internally std::function stores a
pointer to the lambda and a small fixed-size buffer to hold the state of
that lambda. In this case the size of that buffer is sufficient.

After this patch invoking the function is slightly more efficient (one
indirection less), but that was not the goal of this patch. The goal was
to reduce the number of heap allocations, and that's achieved: before we
had many small allocations, now we have one large buffer for the vector
(or a few blocks when you take re-allocating the vector into account).

The total memory usage before/after is roughly the same (depending on
the overhead in heap-administration).