[2/45] LedStatus: unique_ptr -> by value
LedStatus contains an array of 'ReadOnlySetting' objects. This class has
no default constructor, so a 'traditional' array of these objects is not
possible. Instead the code had an array of unique_ptr<ReadOnlySetting>
objects. At first this array is initialized with nullptr's. Later we
replace these with pointers to heap-allocated objects.

This patch introduces the 'generate_array' helper function. This allows
to generate a std::array (not a C-style array) by invoking a
user-provided callback function to initialize each element. The callback
receives the index as parameter and must return the value for that
position. One important implementation detail is that the resulting
array is never default initialized (and then overwritten). Instead the
array is directly initialized with the final values.

So this both reduces heap-allocations and removes an indirection.