[4/45] MSXS1985: unique_ptr -> by value
The MSXS1985 class always initializes a SRAM object during it's
constructor. Nevertheless the SRAM was stored by unique_ptr. This patch
changes that to store by value.

The 'problem' with store-SRAM-by-value was that SRAM has no default
constructor. That means it must be initialized in the MSXS1985
constructor's initializer list. Though depending on some condition
(related to backwards compatibility) we must call one of two different
(overloaded) SRAM constructors. One way to achieve this is like this:

   : sram(<condition> ? SRAM(a, b, c)
                      : SRAM(a, b, c, d))

In older C++ versions this would create a (temporary) SRAM object and
then copy (or move) this to the 'sram' member variable. But SRAM is not
copyable nor movable. To work around this we stored SRAM in a unique_ptr
instead of by-value.

C++ compilers have always been allowed to elide this copy (or move).
That is instead of constructing a temporary and then copying (moving)
it. It is allowed to directly construct the object in the final
location. Even if this copy (or move) has side effects (e.g. log some
message).

Before c++17(?) this return-value-optimization (RVO) was allowed, but
the class was still required to be copyable (movable). Starting from
c++17 RVO is mandatory and the class is no longer required to be
copyable (movable). So now the above construction does work.