[16/45] IRQHelper: unique_ptr -> enum
We have a 'IRQSource' class that helps with generating a IRQ signal. It
offers functionality like keeping track of whether the signal was
already active or not and offers debug functionality like setting
breakpoints on (de)activation of the IRQ source.

'IRQSource' is for 'regular' usage. We also have 'OptionalIRQ' it's
similar but it can dynamically be configured to generate a NMI instead
of a IRQ or it can be configure to generate neither (as-if the IRQ pin
is not connected to the CPU). So 'OptionalIRQ' is more flexible but it's
a tiny bit slower than 'IRQSource'.

This patch tweaks the implementation of 'OptionalIRQ. Before it was
implemented via a small class hierarchy: we had a 'IRQSink' base class
and 3 possible implementations (which do either nothing, lower/raise the
IRQ or lower/raise the NMI).

In earlier patches we already saw that in some cases it can be
beneficial to replace a class hierarchy with a std::variant. That was my
first attempt for this patch. This eliminates the need for heap
allocation and replaces virtual-dispatch with std::visit() (both are
basically the same, both jump-table based).

But in this specific case, and with the current state of the compiler's
optimizers it's possible to do better. We only have 3 alternatives in
the variant, and the state of the alternatives is all the same. This
means we can easily move the state outside the alternatives and then
basically our variant is just an index. std::visit() then becomes a
switch() on the index.

I verified that current compilers generate better code for an actual
switch than for std::visit() (with few alternatives). So in the final
version for this patch I replaced the class hierarchy with an
enum. And instead of having (two) virtual functions we have switch
statements.