[14/45] Return MSXtar by value (was unique_ptr)
This is similar but not quite the same as the previous few patches. The
structure of the 'getMSXtar()' factory function is:

    MSXtar getMSXtar(...) {
        MSXtar result(...);
        result.configure(...);
        return result;
    }

The compiler is allowed to elide the copy of 'result' into the final
location (= some variable at the caller side), but this is not required.
Technically this is a case of NRVO (Named Return Value Optimization)
while the previous patches relied on regular RVO. NRVO is allowed but
not required, and thus a copy/move-constructor is still required but
MSXtar has neither. For more technical details on RVO/NRVO see:
    https://en.cppreference.com/w/cpp/language/copy_elision

This patch adds a move-constructor in MSXtar. Optimizing compilers will
not use it, but (for now) it's still required.