[11/45] MSXPSG: unique_ptr -> by value
The order of initialization of the 'ay8910' member is critical. It needs
to be initialized after the 'ports[]' member. Though because we
initialized 'ports[]' in the constructor body (not in the initializer
list) we could also only initialize 'ay8910' in the constructor body.
And because AY8910 has no default constructor (and is not assignable)
this meant we could not store 'ay8910' by value (instead it was stored
via unique_ptr).

This patch uses the (new) generate_array() helper function to move
initialization of 'ports[]' to the initializer list. And that allows to
also move 'ay8910' there and store it by-value.