[17/45] CommandLineParser: unique_ptr -> optional
In the previous patches in this series we often replaced a unique_ptr to
an object with storing that object by-value. Unfortunately that's not
always possible:
a) There might be strict requirements on the order in which the object(s)
   get initialized. And this might be incompatible with initializing the
   (by-value) object in the constructors initializer list.
b) It might be that we don't (always) need to initialize the object (so
   the unique_ptr remains nullptr).
c) It might be that the unique_ptr points to a base type and we
   initialize it with an object of a derived type.

In case c) the unique_ptr can sometimes be eliminated by using a
std::variant. Though in this patch series we don't consider this
possibility (maybe in future followup patches).

For cases a) and b) it's sometimes possible to replace std::unique_ptr
with std::optional.

You can think of std::optional<T> as:
* A memory buffer suitable to store an object of type 'T' (a buffer with
  the right size and the right alignment).
* Plus a boolean indicating whether:
  - The memory buffer is uninitialized.
  - Or the buffer is initialized with an actual 'T' object.

When we compare std::unique_ptr with std::optional:
* 'unique_ptr' heap-allocates the memory but only when needed. While
  'optional' has the memory pre-allocated (in the same memory block as
  the containing object).
* 'unique_ptr' represents the non-initialized case via 'nullptr'.
  'optional' uses one of the two boolean values for this.
* The total memory usage is:
  - 8 bytes (the size of a pointer) for a nullptr unique_ptr.
  - 8 + sizeof(T) bytes (+ heap-overhead) for an initialized unique_ptr.
  - 1 + sizeof(T) bytes (+ padding) for a std::optional, either
    initialized or not.

Going back to case a) and b) from the top.
* Case a), when we will (almost) always initialize the object (but maybe
  not immediately in the constructor initializer list) then it's always
  a good idea to replace unique_ptr with optional.
* Case b), when we don't always have an object (so when it's truly
  "optional") but the size of the object is small, it may still be a
  good idea to replace unique_ptr with optional.
* But when the object is not always needed and when it's big then it's
  probably not a good idea to use std::optional.
So we have a trade-off between heap-allocation (and an indirection on
use) versus possible over-allocation (and no indirection on use).

--

This specific patch replaces a unique_ptr<FileTypeCategoryInfoTopic>
with optional<FileTypeCategoryInfoTopic> in the CommandLineParser class. The
FileTypeCategoryInfoTopic object is always needed, but it cannot be
initialized too soon. So it's case a) from above.

All later patches in this series make similar replacements for either
case a) or b).