Cleanup 'using std::xxx' that had no or only a few uses
Often these are left over after an earlier code refactoring. For
example, in the past we had a function like:
    void f(vector<T>)
later that was changed to:
    void f(span<T>)
but we forgot to remove the 'using std::vector'.

Or sometimes we had a 'using std::xxx' that was only used in a few
places. IMHO it's then cleaner to not have the 'using' statement.

This patch netto removes almost 300 lines of code.