2e: Store setting-vales in hash_map instead of XMLElement
Before this patch the SettingsConfig class had an XMLElement member
variable that was constantly kept in sync with how the 'settings.xml'
file would look like if it were to be written to disk at that point in
time.

In particular that member contains something like
    <setting id="speed">110</speed>
for every setting with a non-default value. Each time a setting is
changed we look it up in this structure (linear search) and update the
value. This patch replaces that with a hash_map which can be update more
efficiently.

Moreover this allows to remove several modifying operations from the
XMLElement class. And that's important for future patches in this series
where we change the internal XMLElement representation (which makes a
trade-off: better for mostly read-only data, worse for often changing
data).