Revert "[20/45] MSXMatsushita: unique_ptr -> optional"
This reverts commit 286e81c6301b9691efec69ddf53a78fb90d08a91.
Because it doesn't compile with msvc.

I could reproduce the symptom with this small test:
    https://godbolt.org/z/79joa36qE
This compiles fine with gcc/clang, but not with msvc. I *think* it
should compile with c++17, though I'm not 100% sure.

It's only a small tweak, so we don't miss too much by reverting this
patch.