2g: Replace XMLLoader in SettingsConfig
This patch changes how 'settings.xml' is loaded in 'SettingsConfig'.
Previous patches already eliminated the use of 'XMLElement' for the
internal storage in 'SettingsConfig'. But it was still used for loading,
and then immediately converted to the internal format. Now we directly
produce the internal format.

Although this is more efficient, the motivation for this change is
actually to further simplify uses of XMLElement and XMLLoader (one of
the next patches will fully remove XMLLoader).