2a: Factor out XMLEscape code + unittest
XML-escaping is for example replacing '<' with '&lt;'.

This patch factors out the code. Before it was a (static) member
function of XMLElement. Now it is a free function. This patch also
adds a unittest for this new function.

This patch also makes some tweaks to the algorithm. These tweaks are
motivated by a future use case where we write an escaped XML string
directly to a file (so without first constructing the escaped string in
memory), and we want to minimize the amount of write() filesystem calls.
That is we don't want to write() one character at a time.