2b: Add XMLOutputStream + unittest
When creating a savestate/replay, we first build an in-memory
representation of an XML file. Then the only thing we do with this is
convert it to a string and write that string to a (gzipped) file. With
only a little extra care it's possible to directly write the result to a
file without ever building the full in-memory representation.

This patch adds a a helper class for this, plus unittest. See comments
in the code for more details.

This patch adds the helper but doesn't use it yet.