2c: Use XMLOutoutStream for savestate
This patch uses the newly added XMLOutputStream class to optimize
writing savestate/replay files: instead of building an in-memory XML
file representation, we directly write the result to a file, see
previous patch for more details.

This greatly optimizes writing out savestates (saves ~1000 small
memory-block allocations). Though that was not the main motivation for
this patch. Instead the motivation is to simplify the usage-patterns of
the current XMLElement class.

Together with the next patch in this series, this eliminates the need to
dump XMLElement to string and (maybe more important) reduces the number
of variations in how these structures are build. Less variations make it
easier to later refactor the XMLElement implementation.