2d: save settings.xml via XMLOutputStream
Similar to the previous patch. This greatly optimizes writing out the
'settings.xml' file. Though performance of this operation is not
critical at all. Instead the motivation for this patch is to remove the
last usage of the XMLElement::dump() method. And maybe more important:
it's a preparation for the next patch which will remove several
modifying operations from XMLElement.