3a: Added zero-terminate option in rapidsax
Extend our rapidsax XML parser (based on
http://rapidxml.sourceforge.net/) to optionally zero-terminate strings.

This parser reads the XML file in memory and then modifies this memory
block in-place. For example undo XML escaping (e.g. "&lt;" -> "<"). Now
it can also (optionally) place a zero-character after the strings (tag-
and attribute-names, data and attribute-values, ...).

Without zero-termination you can use a std::string_view (basically a
pointer + length) to refer into the buffer. With zero-termination a
pointer is sufficient (but retrieving the length requires some
calculation). So this patch allows callers to make this tradeoff (and in
a later patch the refactored XMLElement class will choose for the
zero-terminate option).