3b: Added new XMLElement implementation (not yet used)
This patch adds a refactored version (almost a fully rewritten version)
of XMLElement. The new class is currently called 'NewXMLElement', but
that's only temporary. Once all users are converted, we will remove the
old version and rename 'NewXMLElement' to 'XMLElement'. (Doing
everything at once would be a too big patch).

The old XMLElement class has a pretty straight-forward representation:
    std::string name;
    std::string data;
    std::vector<XMLElement> children;
    std::vector<std::pair<std::string, std::string>> attributes;
So it owns all strings (that implies it makes a copy of the string read
from the XML file). And it stores all collections in vectors.

The new XMLElement class uses this internal representation:
    const char* name;
    const char* data;
    XMLElement* firstChild;
    XMLElement* nextSibling;
    XMLAttribute* firstAttribute;
with XMLAttribute:
    const char* name;
    const char* value;
    XMLAttribute* nextAttribute = nullptr;

The main differences are:
- This only refers to (zero-terminated) strings, but doesn't own that
  string-memory. This allows to simply point to the string in the
  xml-file-buffer instead of having to copy it.
- The vectors are replaced with single-linked lists. Very often these
  vectors contained relatively few elements. So we had many small
  heap-allocations.
- The sizeof old vs new XMLElement is almost 3x smaller (112 -> 40
  bytes), even without taking the storage of the vector-buffers into
  account.

There are some trade-offs:
- Because XMLElement refers to strings in the xml-file-buffer we must
  keep that buffer around for the lifetime of this structure.
- When we want to modify some string, e.g. to a value that's not present
  in the original file, then we must also manage ownership of that new
  string-memory.
- We must still allocate and manage ownership of the XMLElement and
  XMLAttribute objects (before these were owned by the vectors). Because
  without extra care these could also cause many small heap allocations.
- Storing collections is a single-linked list is less flexible than
  storing them in a vector. Iterating over a linked-list has the
  potential to be bad for cache-locality, but see note about allocator
  below.

To address these we introduce one extra class: XMLDocument. With
representation:
    XMLElement* root;
    MemBuffer<char> buf;
    std::pmr::monotonic_buffer_resource allocator;
So XMLDocument refers to the root element. It also stores the
xml-file-buffer. And finally it has a monotonic allocator. This
allocator is used for the XMLElement and XMLAttribute objects, but it
also can store extra strings e.g. in case you want to modify some
strings.

A monotonic allocator basically is:
- allocation: grab the next chunk from a large block (or allocate a
  fresh block when space has run out).
- deallocation: nothing at all. In other words: freed memory will NOT be
  reused.

While loading an XML file, we only need to allocate memory (never free
it). Thus using a monotonic allocator is a good fit. Also because
allocated blocks are guaranteed to be adjacent in memory. This implies
that XMLElement/XMLAttribute objects are laid-out in depth-first order
in memory. And thus iterating over linked-lists will in practice still
be cache-friendly. When we're done using the XMLDocument, the monotonic
allocator can be dropped as a whole without the need to free all
individual XMLElement/XMLAttribute objects.

If we want to modify a XMLDocument after it has been loaded from disk,
we may need to allocate extra strings and/or XMLElement/XMLAttribute
objects. If these new strings/objects replace old ones, then those old
ones will NOT be freed (or not until the whole document is freed). This
makes the new XMLElement infrastructure less suited for heavy-modify
situations. (This was the case for the XMLElement in SettingsConfig, but
earlier patches in this series already refactored that).

In some cases we don't load an XML file from disk, but build an
XML(-fragment) from scratch. In that case the file-buffer will remain
unused and everything is allocated from the monotonic allocator.

Some of these fragments remain valid for the whole duration of openMSX
(before these were 'static XMLElement' variables). To manage ownership
of these static fragments we have a singleton-like static XMLDocument.
It is possible to refer from other XMLDocument to such static fragments
(so cross document references). As long as the lifetimes are obeyed
everything is fine.