3c: load savestate via new XMLElement
Load savestate and replay files via the new XMLElement infrastructure.
Both new and old have roughly the same public API, so this is a
relatively small patch.

Loading a (freshly) saved savestate traverses the XML tree exactly in
depth-first order. And with that property it's not required to first
read the whole XML file in-memory (we could stream the part we need
just-in-time). However when loading older versions we allow to reorder
the (direct) child tags. Or sometimes we have to traverse small sub
trees in a different order to reassemble information that's now (in the
new version) stored differently. So having the full tree in-memory is
still useful.