3d: use NewXMLElement in HardwareConfig, remove old, remove XMLLoader
This patch is the combination of 4 things:
 1) Use NewXMLElement in HardwareConfig and related code.
 2) Remove the (now no longer used) old XMLElement code.
 3) Remove the (now no longer used) XMLLoader code.
 4) Rename NewXMLElement to XMLElement.
Though I've combined these in a single patch because otherwise step 1
would make a lot of renames that would be undone in step 4. And the diff
for this whole rename stuff would be larger than this patch itself.

HardwareConfig is a mostly-read-only structure. It contains the XML
machine/extension descriptions. It's not 100% read-only because e.g.
after instantiation we fill-in concrete slot numbers for slots specified
as "any" in extensions. And e.g. we add a concrete 'resolvedFilename'
tag for files that were searched by 'sha1'.

But still because HardwareConfifg is mostly-ready-only and because the
query-part of the old/new XMLElement API is largly the same, this patch
is relatively small. The creation/modify-part of the API does differ
more, so most of this patch touches that part of the code.

--

As a bonus this patch also simplifies how savestate/replay files look:
Statestates are (gzipped) XML files, somewhere inside they contain the
machine config which is also an XML file. Before that machine config
part was dumped as something like:

    <config>
      <name>machineConfig</name>
      <data/>
      <attributes/>
      <children>
        <item>
          <name>stuff</name>
          <data>someData</data>
          <attributes>
            <item>
              <first>id</first>
              <second>1</second>
            </item>
          </attributes>
          <children/>
        </item
      </children>
    </config>

Now we directly embed the machine-config-xml inside the savestate-xml,
and the stuff above just looks like

    <config>
      <machineConfig>
        <stuff id="1">someData</stuff>
      </machineConfig>
    </config>

The only reason why before it was dumped in this complex format was that
this was easy to implement in the current serialization framework (see
the backward-compatibility code OldXMLElement::serialize()). In other
words: the old format was accidental.