Made several base classes non-copyable
And non-copyable implies non-movable.

In particular classes that have some associated
    registerXX(BaseClass&)
function.

The reason for this is that the registerXX() function will internally
store the address of the given object. But then that pointer better
remains valid. In other words: the object better not move to a different
location in memory (after being registered).

This change was triggered by this bug report on MRC:
    https://msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=57
    2nd post on the page, by 'ren', 24-08-2021, 01:04
I can't reproduce this bug on Linux. I suspect it's because of a
difference between gcc/clang and msvc, the way how both compilers handle
c++17 copy-elision (bug in msvc??).

If my suspicion is correct, then after this patch there will be
compilation errors on msvc (already better than wrong behavior at
runtime). If that's the case then I'll fix those in a follow-up patch.