Implement our own version of std::pmr::monotonic_buffer_resource + unittest
Because libc++ doesn't have this class yet, even though it's part of
c++17.

(Before this patch) this broke the android build, and probably also the
macOSX build.

This adds a very minimal re-implementation of monotonic_buffer_resource,
just enough for what we currently need. The standard API is more
extensive. Because of this difference I intentionally did not pick the
same name (I picked monotonic_allocator).