Fix error on 'type \r ; reverse savereplay'
To reproduce execute a command like this in the console:
    type symg9k\r ; reverse savereplay bla
And this would trigger an error like
    could not write "bla.omr"

The problem occurred due to an 'unfortunate' API choice in the gzwrite()
function (part of zlib). The return value of this function is
  * on success: the number of written bytes (uncompressed)
  * on failure: zero
Though if you call gzwrite() to write zero bytes, the function
(correctly) writes zero bytes (so does nothing) and then returns the
number of written bytes (zero). But unfortunately that return value is
also the way to signal an error. So before this patch, openMSX
interpreted this as "some error has occurred", and this lead to the
"could not write ..." error message.

This patch changes the check from:
    if (gzwrite(...) == 0) {
to:
    if ((gzwrite(...) == 0) && (len != 0)) {

But why would you ever want to write zero bytes? Let's first see why it
occurs and then reason about if it's a good idea or not.

This zero-length-write was triggered from the new XML-escaping code
(e.g. replace '<' with '&lt;'). Before this routine produced a newly
allocated output string. Though the majority of the strings don't need
any escaping and then we make an unnecessary allocation/copy. In the new
routine we scan the input for the first to-be-escaped character. We then
'output' part of the string before this character as-is, followed by the
escaped character. Then we continue with the rest of the input. (What
'output' does is still configurable, but in this context it means send
it to gzwrite()).

In some cases, e.g. when the string ends with a to-be-escaped character
(like in the 'type' command above), some of the output chunks can be
empty. We have a unittest for this function and it does test these
corner cases (and those tests do pass). However the unittest is
appending the chunks in a memory buffer. And that (already) deals well
with the case of appending a zero-length-chunk (gzwrite() did not).

This patch made gzwrite() with a zero-length buffer just work. But
should we instead detect this in XMLEscape() and avoid writing
zero-length chunks in the first place?
* From a robustness point of view it makes sense to make gzwrite() deal
  with length=zero. Who knows what other callers (now or in the future)
  can produce zero-length-chunks.
The question then becomes: should we in addition detect zero-length
chunks in XMLEscape()?
* That's a valid option, but it's not my preference. I personally
  prefer code with less special cases. That means not having a special
  case for length=0.
But isn't this bad from a performance point of view? IOW why call a
function that we know will do nothing?
* It depends. Calling gzwrite() with length=0 is fairly quick (I checked
  the implementation). We could also check ourselves and then don't do
  the call at all. But this test itself also takes time, and most of the
  time there won't be any empty chunks (and then the extra check only
  adds overhead). So really it depends.

--

This patch also fixes a secondary problem. After the problem above
triggers the code throws an exception. During stack-unwinding we cleanup
stack objects, among them the XmlOutputArchive object. During
destruction of this object we write the final closing '</serial>' tag
and close the file. Though because not the whole XML file was written,
this closing tag is not properly paired with the corresponding opening
tag. And (only in a debug build) we detect this mismatch and trigger an
assert. So for end-users this secondary problem won't occur, but it's
better to fix it anyway.