Fix possible crash when switching V9990 display modes
We got a (for me difficult to reproduce .. it's very timing sensitive)
crash in SymbOS (for gfx9000) when it's switching from 512 to 384
(overscan) pixel wide display mode.

The symptom was:
  openmsx: src/video/v9990/V9990PxConverter.cc:304: void openmsx::V9990P1Converter<Pixel>::convertLine(Pixel*, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, bool) [with Pixel = unsigned int]: Assertion `displayWidth <= 256' failed.

After some debugging we found that displayWidth=288. Digging deeper that
value 288 comes from the overscan horizontal timings. But overscan and
P1-mode cannot go together.

This inconsistency was caused by the V9990-core and the V9990-renderer
not being 'in-sync' about the active display mode.

More in detail: on a mode change (actually on any state change), we
first render the screen up-to the current moment in time and only then
make the actual change. Future renders will then use the new state.

Though we had a bug that (wrongly):
* first set the new display mode in V9990-core
* then renders up-to the current moment in time
* then sets the new display mode in V9990-renderer
During rendering the renderer queries stuff from V9990-core, some of
these queries depend on the mode. But V9990-core was (wrongly) already
using the new mode, while V9990-renderer was (correctly) still using the
old mode. Hence the inconsistency.